moving run command in self file
