+++ packages/cli/src/build.ts
@@ -1,0 +1,28 @@
+import { GENAI_EXT, host, parseProject } from "genaiscript-core"
+
+export async function buildProject(options?: {
+    toolFiles?: string[]
+    specFiles?: string[]
+    toolsPath?: string
+    specsPath?: string
+}) {
+    const {
+        toolFiles,
+        specFiles,
+        toolsPath = "**/*" + GENAI_EXT,
+        specsPath = "**/*.gpspec.md",
+    } = options || {}
+
+    const gpspecFiles = specFiles?.length
+        ? specFiles
+        : await host.findFiles(specsPath)
+    const scriptFiles = toolFiles?.length
+        ? toolFiles
+        : await host.findFiles(toolsPath)
+
+    const newProject = await parseProject({
+        gpspecFiles,
+        scriptFiles,
+    })
+    return newProject
+}
