+++ packages/core/src/index.ts
@@ -55,3 +55,4 @@ export * from "./xml"
 export * from "./aici"
 export * from "./file"
 export * from "./fetch"
+export * from "./edits"
